[TM-2] Wrap annotations with newtype
Problem: annotations are just `Maybe Text` and could be
and each of them could be passed instead of other ones by mistake.

Solution: created `Annotation` newtype which accepts `tag` type parameter,
which can be either `TypeTag` or `FieldTag` or `VarTag`. Also in related types `Note`
substring was replaced with `Ann` (stands for annotation).