[TM-13] Add interpreter infrastructure
Problem: there is a stub for Michelson interpreter, but there is no code
which uses it. It should read data from CLI, maintain some state and pass it
to Michelson interpreter.
Solution:
1. Add `GState` with emulated global blockchain state and `TxData` with all
necessary data about a transaction.
2. Define serialization of `GState` (in JSON).
3. Add CLI for interpreter.
4. Add interpreter engine which calls `michelsonInterpreter`.
This is only a skeleton of the final solution and there are many TODOs which
will be addressed separately.