[TM-2] Add support of annotations for advanced types
Problem: advanced type representation of Michelson language
(strictly-typed GADT-based) lacked support of type and field
annotations.

Solution: add support for annotations as value-level structure being
carried and updated as part of stack type while doing the type check.

Note: no convenient support for construction of annotated Michelson
values is given yet, all values of `Instr op inp out` are to be
interpreted as annotations-free typed values.
This should not be an obstacle for practical usage of the types though.