[TM-13] Add origination command
Problem: before calling a contract it should be originated. Currently
we don't have a way to originate a contract (except manual modification
of JSON file). We need a command for it.
Solution: replace `interpreter` with `originateContract` and `runContract`
functions. Change `InterpreterOp` into more correct type: a contract can be
originated or called by transferring tokens to it. Now `run` command
originates the contract first and then calls it pretending that a transaction
is sent to it.