[TM-22] Make `initGState` empty
Problem: documentation for `initGState` states that it's empty, but it's not.
In fact we don't need it to be non-empty, it was needed earlier when `run`
command didn't originate the contract, but now it does and we have a separate
origination command, so empty state is fine.
Solution: make `initGState` empty.