[TM-18] Make current time and steps limit CLI options
Problem: ContractEnv contains current time (ceNow) and maximal steps
(ceMaxSteps), but both values are hardcoded in Morley.Runtime. We
should allow them to be passed in as CLI options and have default
values if they are not passed in.

Solution: add 'time' library, add implmentation for getCurrentTime,
add MaxStep and Now options. MaxStep has a default value,
now is optionally parsed, if it is Nothing then we get the current time.