[TM-48] Enhance `Timestamp` type
Problem: `Timestamp` is a newtype over `Word64`, which is too
arbitrary. There are existing more specific types to represent
timestamps. The essence of this type is clearer if it wraps
existing timestamp type, also time-specific types are easier to use
with existing time-related functions.
Solution: change internal representation of `Timestamp` to `POSIXTime`,
because timestamps in Michelson are essentially Unix timestamps.
Provide some utility functions.