[TM-17] Make contract's address computation more correct
Problem: we have 'contractAddress' function to compute address of a
contract. However, it returns a constant value, which is also invalid.
Solution: implement address computation in a similar way to what Tezos
reference implementation does based on this stackexchange answer¹.
Since we don't know precise binary format yet, it uses JSON
serialization as the easiest solution. We don't need to be precisely
compatible with Tezos at this point.

¹ https://tezos.stackexchange.com/a/361/342