[TM-20] Slightly refactor Morley interpreter to expose GState updates
Problem: Morley interpreter executes GState updates directly. It
makes it harder to test it, because we don't have list of all
performed operations. We have final GState, but it's less precise
(we have to think about initial GState, for example). We also have
updated storages, but this is not the only possible operation.
Solution: add 'GStateUpdate' type and store list of this type, not
only storage updates.