[TM-20] Add integrational test for stringCaller
Problem: existing testing infrastructure and tests deal only with
isolated contracts, but not with their side effects. We need to test
interactions between contracts as well.
Solution: add infrastructure for integrational testing and a test for
'stringCaller' from the tutorial¹.
Note: it uses untyped representation currently. Maybe we should
reconsider it, however, since existing Morley.Runtime uses untyped
types, I think it's better to change it separately.
We still need untyped representation. For example, in order to compute
contract's address we need to serialize it. It's easier to serialize
untyped representation (this functionality is present already and is
needed anyway). And integrational interpereter doesn't really benefit
from typed types. So I think it's fine to keep using untyped types
there.

¹ https://gitlab.com/camlcase-dev/michelson-tutorial/tree/master/03