[TM-20] Add a simple integrational property test
Problem: it should be possible to write integrational property tests
using our framework, but currently it's not convenient to do so.
Solution: provide two versions of `integrationalTest`, one works with
`Expectation`, the other one works with `Property`. Add a simple
example of using the second one. It's similar to the existing tests,
but calls 'idString' with an arbitrary string, not with a constant one.