Extract `Prelude` module
Problem: intero flycheck is known to break in case of custom Prelude unless
it's imported from another package.

This commit extracts 'Prelude.hs' module to a separate package.

Now the project has two packages with all the ensuing consequences (e.g. one needs to do
`stack build morley` in order to see build progress - if he does not use the Makefile -,
same for tests).

I would rate the inconveniences of not using flycheck as 3/10 (slightly annoying), so if
the current model brings you unpleasanties then let's wait with this PR until we have
to split the project into subprojects for other reasons.