Make `valToOpOrValue` return `Maybe`
Problem: a value may contain lists, thus there might be several
operations within a value. There is so trivial way to get them all,
and we don't care much about that since the only code which considers
'Left' result of this function is 'Buildable' instance

Solution: for now we just return 'Nothing' there.