gitlab-ci: Exclude Docker from the CI pipeline
The previous CI pipeline had a lot of unnecessary weirdness which,
unfortunately, resulted in less-than-desired reliability (TM-70).

* Switch to the `docker` executor as `shell` is insecure
* Do not build Docker images, they were not used for anything anyway
* Cache .stack directories
* Use `haskell` as the base image with GHC version matching LTS
* Get rid of stages to speed up builds, they were not used anyway

Notes:

The previous setup with a sequence of Docker images relied havily on
locality of cache. Introducing an additional runner breaks everything
completely as those images are not properly distributed (see TM-70).

Stages are used for parallelising executions of multiple jobs in complex
pipeline, but in our case we, essentially, have only one task. Having
just one job is better, as the coordinator will not have to submit three
jobs to the runners (which, by the way, may be different for each job).
In our case it is necessary to avoid expensive cache push/pull steps.

GHC from the base image is used instead of a stack-managed one simply
because it is faster and does not have any obvious downsides.

Ideally, we would like to have our caches per-branch, but it is
impossible to configure GitLab CI to fall back to cache from `master` if
there is no cache for the branch being built, so having per-branch
caches would result in very fast builds of commits from the same
branch/MR, but slow initial builds, as no cache would be available at
all. Sad.

Aggressive caching of .stack directories might result in unexpected build
failures, because stack is a piece of garbage, but let’s see.