[TM-75] Add a `transfer` command
Problem: currently the only way to send a transaction is to use
the `run` command. It has at least two drawbacks:
1. It can only send a transaction to a contract, not to a key address.
2. One must pass a contract to it, even if this contract has been
originated already. However, if the contract is already originated,
its address is enough to send a transaction to it. Moreover, one must
make sure that computed address will be the same as they expect.
Solution: implement a `transfer` command as a simple wrapper over
existing functionality to send a transaction to an arbitrary address.