[TM-47] Prohibit creation a contract wrong literal
Problem: one can try to PUSH a contract's address
which doesn't exist and then try to call this contract.
As result, code isn't so typesafe, how it could be.
As well, it diverses with reference implementation.

Solution: pass to typechecker's environment a map
from contract's address to its type parameter and
verify that every PUSH of a contract
is valid, namely, a contract's type in PUSH converges with
a type parameter of a corresponding contract from the map.
Functions which depends on typechecking also takes this map now.
Also, 'typeCheckValImpl' was reimplemented to return an annotation
aming with type.