[TM-78] Define `Eq` instances for typed values
Problem: in order to check whether to typed values are equal, you have
to pattern match on them explicitly or convert them to untyped values
and then compare. It's not convenient comparing to the most natural
`==` operator.
Solution: implement `Eq` instances for typed types. Introduce
`eqParam*` helpers to check equality of types inside GADTs.
For typed `Instr` it's implemented based on conversion to untyped
instr. I think it's definitely better than defining it manually,
because there are too many cases and it's very easy to make a mistake.
It has two drawbacks:
1. Orphan instance. Shouldn't be a big deal as long as we are careful.
It's unlikely that we'll define it somewhere else.
2. Conversion may be expensive. Also shouldn't be a big deal, because
`Eq` is mostly for values and `lambda`s should be quite rare I guess.