[TM-78] Pass less values to `ContractPropValidator`
Problem: `ContractPropValidator` has many arguments, but it's only
used in `contractProp` and these arguments are redundant there, because
they are passed to `contractProp` itself. So the caller of
`contractProp` knows them anyway. They only advantage of this approach
is that we can use the `Testable (a -> prop)` instance, so that all
arguments will be generated automatically and we won't have to mention
them, but existing tests demonstrate that it's a rare case.
Moreover, `ContractPropValidator` validates in terms of typed values,
but it's easier to just use Haskell values.
Solution: remove 3 arguments from `ContractPropValidator`. Put them
into validators' closures, because they are always available before
calling `contractProp` anyway.