[TM-91] Make `{}` parse to specific `ValueNil`
Problem: At parsing stage we does not know which kind of list we parse if it is
empty - is it a plain list, set, map or lambda.

Solution: Typechecker has enough information to distinquish them, so we defer
decisions about empty lists to type check stage via dedicated `ValueNil`
constructor.