[TM-58] Update Michelson.Printer
Problem: Previous version of  was very old and out of sync with changes to the
rest of the project. Also we parse PAIR we treat it as a PAPAIR macro.
When we parse PAIR we treat it as a PAPAIR macro. It's an old hack. Because of
that when we parse a contract with PAIR we get a different contract: PAIR turns
into { PAIR } (a sequence with one instruction).

Solution: Define a RenderDoc typeclass that uses 'wl-pprint-text' to print a
Michelson or Morley contract into a valid contract that Tezos can read. We
currently print the contract as a single line. Add code to handle PAIR
specially.