[TM-78] Slightly tune EnvironmentSpec
Problem: EnvironmentSpec test was able to discover the bug that is
fixed in the previous commit, but the chances of finding this bug
were slow, so we found it with delay.
Solution: decrease range of possible values for `fMaxSteps` and
increase max success to ensure that the test will quickly fail if gas
consumption changes again.