[TM-27] Fail properly
Problem: previously it wasn't clear how to handle fails in the
typechecker properly, so we always printed not the top of the stack
right before `FAILWITH` but top of the stack at the beginning of the
code block. Now we have time to experiment on this and come to some fix.

Solution: extend `SiFail` constructor so that it contains instruction
with arbitrary output stack, and no actual output stack information -
this matches `FAILWITH` scenario perfectly.

Input stack is extracted to be part of outer datatype, since it is a
common part for `:::` and were-`SiFail` constructors; this allows us
avoiding duplication when processing input stacks returned by
type-checker.