[TM-118] Attach source code location to instructions
Problem: We have to enhace error messages produced by
the typechecker. To achieve this goal we have to print precise
location of an ill-typed instruction among with an error message.

Solution: First of all, to tackle this problem we need to
attach line and column where an instruction starts to this
instruction, to make it possible for typechecker to report
the location if the instruction turns out to be ill-typed against
a current stack.
Therefore, 'Michelson.ErrorPos' module is created
and some basic datatypes which related to location are introduced there.
Parsed and Expanded operations are extended by this datatypes and
code has been refactored correspondingly.
At present, introduced location information isn't exploited in
the typechecker because this probably should be done after TM-60
is done (and fine grained error type is introduced).
As an additional artifact, 'Morley.Macro' has been refactored in
a way that reduces number of 'SeqEx` in an expanded contract.
Also, an instruction call stack to some constructors of TCError is addeded.