[TM-68] Add convenient operators for typed instructions
Problem: typed instructions essentially represent functions which
operates on Michelson stack. It's convenient to use some operator
similar to -> as an alias for this data type. Also it's sometimes
convenient to use `&` instead of `':` for type-level list
concatenation (that we use to represent stack type).
Solution: add the aforementioned operators as aliases to existing ones.