[TM-44] Move core instruction parsing into a new module
Problem: Michelson.Parser is a very big module, it's hard to read it.

Solution: the next step towards making all parsing modules small is
to move instruction parsing module into a new module.
It handles parsing of core Michelson instructions and ignores macros
and extensions.
Doing it leads to some cyclic dependencies which we resolve the same
way as for value parser: by passing some parsers as arguments.