[TM-44] Move built-in macro parsing into a new module
Problem: Michelson.Parser is still a bit big module,
it's hard to read it.

Solution: the next step towards making all parsing modules small is
to move macros parsing module into a new module.
It handles parsing of built-in Michelson macros.
Doing it leads to some cyclic dependencies which we resolve the same
way as for value and instruction parsing: by passing some parsers
as arguments.