[TM-68] Resurrect Lorentz
Problem: We had some drafts about Lorentz DSL, which had been buried:(
However, now we want to resurrect it in order to use it in our
standard. Lorentz is intended to make writing Michelson contracts
easier and more convenient.

Solution: Take code from old !53, that was closed and move it to
`src/Lorentz/`. Summary:
* Aliases for types (`Lorentz.Types`)
* Lowercase instructions (`Lorentz.Instr`)
* Value constructors (`Lorentz.Value`)