[TM-121] Add construction of complex datatypes
Problem: we have getters and setters mapping from Haskell to Michelson,
but it's not yet clear how to first construct those datatypes.

Solution: add a `construct` function which accepts instructions,
each creating one of the datatype fields. This function produces a linear
number of instructions and so if fine to use.