[TM-121] Define `if then else` for Lorentz
Problem: plain `if*` instructions are inconvenient to use
when their clauses are big pieces of code - at least it's not clear
how to indent them normally.

Solution: define `if ... then ... else ...` construction for Lorentz.
As soon as Lorentz is stack machine language, condition in `if` stands
for kind of operation we apply to top elements of the current stack.

We take advantage of the fact that when `RebindableSyntax` is enabled
then expression `if a then b else c` is equivalent to `ifThenElse a b c`
diregard the types.