[TM-143] Introduce hpack
Problem: it's annoying to specify all project modules manually,
and there is a lot of duplication in `.cabal` files we carry around
with each added subproject.

Solution: use hpack.

Generated cabal files are very close to the original ones.
The only difference I noticed are

* Absence of `autogen-modules: Paths_morley`, but
[this issue](https://github.com/sol/hpack/issues/303) states that
hpack does this intentionally and this should be safe.

* Cabal version has downgraded from 2.4 to 2.2, but hpack is said
to use the smallest version which is enough to support all required
features, so this should be fine as well.