[TM-136] Remove `extra-source-files` from prelude
Problem: referring files outside of the current directory in
`extra-source-files` section of `.cabal` file is not allowed.

This causes the following warnings:
```
Configuring morley-prelude-0.1.0.4...
Warning: 'extra-source-files: ../CONTRIBUTING.md' is a relative path outside
of the source tree. This will not work when generating a tarball with 'sdist'.
Warning: 'extra-source-files: ../CHANGES.md' is a relative path outside of the
source tree. This will not work when generating a tarball with 'sdist'.
```

Actually, before hpack was introduced, `morley-prelude` didn't have
this section at all.

Solution: remove this section.