[TM-130] Add FA1 description
Problem: we have FA1.mtz contract in the repo and we need to have
its documentation somewhere.
Solution: add relevant documentation to `docs/TZIP/`. We decided
to store TZIPs which have associated contracts in `morley` for a
while. It makes their maintenance easier (they are less likely to
diverge). Once things get more stable we'll move these docs to
the `tzip` repo.