[TM-145] Change approach to uncaught exception handling
Problem: if some exception is uncaught, we catch it in `main` and
unconditionally call `die` with the result of `displayException`.
However, we shouldn't do it for `ExitCode`, because it specifies
exit code of application and `die` always exits with a constant
code (1).
Solution:
1. Ignore uncaught `ExitCode`, rethrow it as is.
2. Offload uncaught exception printing to default exception handler
by wrapping uncaught exceptions into a special newtype.
It also solves issue with print non-ascii characters to stderr when
current locale doesn't support them (this issue can be solved
differently though).
3. Move this logic into a separate function, because now it's less
trivial.