[TM-173] Do not pass --local-bin-path when installing external tools
Problem: when we install `weeder` and `crossref-verifier` we pass
`--local-bin-path`, so that executables are put into specific location.
The only reason we install them is to launch them (unlike `morley`
which we also upload as artifact). Default installation location seems
to be in `$PATH`, so we can just install there and run them. So this
`--local-bin-path` option is not needed and it's better to remove it
to make CI config slightly simpler (otherwise people may wonder why
we pass it). Note that it doesn't spoil caching: built executables are
put into `.stack-work` which is cached and `stack install` simply
copies them if they are built already (it's cheap).
Solution: remove this option. Also unify installation commands.
Call these tools without any prefix, they should be in `$PATH`.