[TM-178] Add lorentz contracts registry
Problem: we need to keep a registry of written Lorentz contracts
which could be printed when necessary.

Solution: add an executable which can dump contracts.

A simple automatic contracts lookup is implemented which traverses
haskell modules and finds all contracts of `contract_<name>` format,
and fills a module of all found declarations like e.g. tasty-discovery
does it.