[TM-120] Enhance instructions working with sum types
Problem: for modularity of contracts we'd like to be able to merge
store templates into one big template. First step to that is to
enhance instructions like `instrWrap` and `instrUnwrapUnsafe` to
search for fields transitively.
So if there is `data A = A B | B C` where `B` and `C` sum types, we'd
like to be able to easily wrap parts of `B` and `C` into `A` and vice
versa.
Solution:
1. Extend 'GLookupNamed' to go deeper if it encounters a type that we
can recursively inspect.
2. Add 'S' (straight) branch to support types which wrap other types
into a single constructor.
3. Provide `GInstrUnwrap` and `GInstrWrap` instances for new cases.
4. Add examples with compound types.