[TM-120] Add a prototype of DS token
Problem: we need to implement DS protocol one part of which is a token
and it's not implemented.

Introduction:
Since token needs to interact with other parts (e. g. registry
service) we need some architecture to support it.
The easiest approach is to put all logic into one Michelson
contract. But we want our code to be maintainable and don't want
to put all original code into one place.

Solution:
1. Add `DSProtocol` contract to `lorentz-contracts`. Its
documentation describes developed architecture.
2. Add a prototype of DS token, mostly for illustration purposes.
It implements only Mint and Burn as well as getters for balance and
total supply. It's supposed to be extended later (TM-184).
3. Update registry contract (now it probably can't be called
a contract) and tests accordingly.