[TM-179] Allow specifying source in interpreter
Problem: in tests we often need to make a transfer like if it was
initiated by a given desired address.

Solution: extend `TransferOp` to accept source address optionally;
this may be useful for the morley executable some day as well for
testing purposes.
Add `withSource` function to be used as building block of integrational
testing scenarious.