[TM-172] Avoid exponential growth in `Type` gen
Problem: generator for `Type` is recorsive, and as soon as we have
`Pair`, `Or` and other stuff with two arguments and producing
a leafe in type tree (unit, primitive or operation) is not very
probable, we can face generation of exponentially-sized data.

Solution: when generating types for `Pair`, `Or` or `Lambda`,
produce leafs more often than usual.