[TM-191] Add DS Protocol documentation
Problem: we now have a DS protocol contract which is quite complex
already and will be more complex later. There is some documentation
in our Haskell code, but we also want to make it easy to use this
contract for people who are not familiar with our Haskell code.
So we need to write some documentation which will help people use
our contract.
Solution: add a document which describes the contract in general,
its parameter and storage, provides instructions for transpiling to
Michelson and describes some entry points.