[TM-216] Prohibit setting pause to its current value
Problem: according to the Ethereum implementation it shouldn't
be possible to SetPause to the same value twice in a row.
And it doesn't make sense to permit it, because it will simply put
a redundant transaction to blockchain.
Solution: fail if someone tries to do it. Add tests for Pausable.