[TM-190] Apply migrations in upgradable contracts
Problem: Storage state was not updated so it was hard to roll out
upgrades that change the storage semantics (UStorage values and
their corresponding types).

Solution: Migrations are lambdas passed to the contracts that are
executed during an upgrade. They are aimed to solve this problem.
However, in order to solve it in a type-safe way, these migrations
probably have to be autogenerated.