[TM-190] Support BigMap in upgradable storage
Problem: One could not set and get values of type BigMap in UStorage
due to the fact that big maps can't be packed.

Solution: Change the signature of UStorage to accept ByteString as
the key parameter and modify accessors to use `pack ((), MText)` for
field keys. Also provide functions to access big maps in storage
using `pack (MText, kt)` where the former is the big map name and
the latter is big map key.