[TM-190] Add UpgradableUnsafeLedger contract
Problem: There should be an example of near real-life problem solved by upgradable
contracts. There should also be an example of using BigMaps in upgradable storage.

Solution: Provide an UpgradableUnsafeLedger based on the existing UnsafeLedger
contract. Use a buggy implementation in the first version and show how to fix
the bug in the second version.