[TM-201] Enforce equal types in COMPARE
Problem: typed COMPARE instruction can work with any `n` and `m`
types for which `ArithOp Compare` is defined. We know that in
Michelson this instruction is defined only for pairs of equal types,
you can check it if you look at `ArithOp Compare` instances.
It means that in definition of `COMPARE` we can require that `n` and
`m` are the same. We can use this knowledge in some places, for example
to optimize `PUSH nat 0; COMPARE; EQ`.
Solution: remove `m` parameter, use `n` as type of both arguments,
update all other code accordingly.