[TM-201] Add optimizer
Problem: many Michelson contracts are not optimal in a sense that
they can be replaced with a contract which has exactly the same
semantics and interface, but consumes less gas.
We'd like to automatically optimize contracts by applying some rules.
Solution: add a module which optimizes typed `Instr` and CLI for it.