[TM-201] Refactor and improve optimizer
Problem: current version of optimizer does not account for:
1. Instructions can be stored in other instructions (like `IF`) and
they should be optimized as well.
2. Rules like `SWAP; SWAP` can be applied only if there are other
instructions after `SWAP; SWAP`, but not to exactly this sequence
(of two instructions).
3. First argument of 'Seq' can be 'Seq' as well, so instructions can
be represented as arbitrary trees.

Solution: fixing these issues straightforwardly makes optimizer a bit
messy, so some refactoring was performed. Now each rewrite rule is
represented as a standalone function and we provide generic functions
which work with rules. In particular there is a function which applies
given rule to deeper instructions (inside `IF` and similar
instructions). It solves the first issue. The second issue is resolved
by adding `Nop` instruction to the end of each sequence and erasing it
later. In order to resolve the third issue I've added 'linearizeRight'.
It should be improved (see TODO), but let's do it later.