[TM-190] Use UParam in UpgradeableUnsafeLedger
Problem: The old interface is not type-safe, thus it is easy to
make a trivial mistake while calling the contract. The solution
offerred by @martoon in TM-246 provides a type-safe wrapper over
(MText, ByteString), and we should use it in examples.

Solution: Add UParam to the implementation of UpgradeableUnsafeLedger,
incidentally deliver some modifications in common upgradeable types.