[TM-100] Increase indentation for nested blocks
Problem: Currently nested blocks does not have any extra indentation
apart from the single curly brace that surrounds the block, making it
hard to identify the code block a given instruction belongs to.

Solution: Enclose the content of the nested block in spaces so that
it gets extra indentation in addition to the enclosing curly brace.