[TM-100] Add pretty printer tests
Problem: Add tests for pretty printer formatting behavior.

Solution: Add sample contracts with their prettified version to
`contracts/pretty` folder and add tests that calls pretty printer over
them and compare the output with their prettified version.  This is a
bit fragile, as *any* change in pretty printer behavior will break the
tests.  But since it is hard to specify and evaluate a prettified
output, this method seems to provide the best bang for the buck.