[TM-190] Ensure consistent spelling of Upgradeable
Problem: There were no certainty on how we spelled Upgrad(e)able
in this repository. In the documentation and UStore we wrote
"Upgradeable" while in the contracts we used "Upgradable". Basic
googling showed that both variants were perfectly valid both in
AmE and BE.

Solution: Use Upgradeable (with "e") everywhere.