[TM-197] Accept single instruction blocks without braces
Problem: The parser right now requires braces around sequences/blocks
even when the said sequences/blocks contain just a single instruction.
The problem is how to remove this source of user unfriendliness.

Solution: If the parser cannot find an opening brace, then make it look
for a single instruction and make a sequence out of it. This is
trivially achieved by adding an altercate branch to the sequences
parser. But this caused the parser to hang at the end of the input.
This was found to happen because the instructor parser and the sequence
parser kept failing, at the end, but just kept calling each other
without moving the parser forward. As a work around, a `lookAhead` call
was added to the sequence parser, which would look ahead for any valid
`letterChar` before starting the next recursive call to the instruction
parser.

Tests: The sample contract from the issue TM-197 was added as a file to
the `contracts` folder, so that it would be included in the parser test.

[TM-197] Fix parser loop in case of bad content at the end of sequence

Problem: The `lookahead` function that was supposed to prevent the
parser loop only looked for a valid `letterChar`. But this would mean
that the loop would still occur if there was some un-parsable content at
the end of the block. This was found to be the case on testing.

Solution: Move the lookAhead to the single op parser, to the choice
alternative that tries to parse naked sequences
(sequences that are not preceded by an instruction). And here make it
look for the '{' character, which was already required by the existing
sequence parser (Michelson.Parser.Instr.ops'). So this basically make
the braces optional for sequences that is part of instruction, but
retains it mandatory for naked sequences.

Tests: Added two tests. One that checks parsing of a DIP instruction
with a single instruction sequence. And one that checks for exception
on parsing a sequence with un-parsable content at the end.