[TM-149] Get rid of contractResShouldBe
Problem: we have `contractResShouldBe` and `validateStorageIs`
functions in interpreter tests. Their logic is essentially the same.
Solution: remove `contractResShouldBe`, use `contractProp` and
`validateStorageIs` in places where it was used. `validateStorageIs`
now uses functions and types from HUnit, because we are slowly getting
rid of `hspec-expectations` and because `testCase` actually expects
`Assertion`.