[TM-274] Update ManagedLedger
Problem: there are some changes¹ in FA1.2 and our implementation
of ManagedLedger does not account for them.
Specifically:
1. FA1.2 now has less strict requirements on parameter type, so
Managed Ledger can be somewhat simplified (no SaneParameter stuff).
2. At the same time FA1.2 has a stricter variation to be usable in
Athens. We do not have an implementation of it.

Solution:
1. Now there are three contracts:
* ManagedLedger implementation for Babylon.
* ManagedLedger implementation for Athens using the approach
described in FA1.2.
* Proxy for the Athens implementation as described in FA1.2.
2. There are two new modules with code shared by these contracts.
3. Tests are updated: their logic is generalized and we instantiate
them with each contract (only some tests are suitable for Proxy).
4. Some tags (which use `named`) are updated to be consistent with
FA1.2.
5. Previously we allowed admin to make arbitrary transfers without
checking approval. This logic was removed to make implementation
simpler.

¹ https://gitlab.com/serokell/morley/tzip/merge_requests/1